+++ continuousintegration/orchestration/tests/changeStorageClassInCatalogue.sh
@@ -29,7 +29,6 @@ fi
 
 EOSINSTANCE=ctaeos
 NEW_STORAGE_CLASS_NAME=newStorageClassName
-TMP_DIR=$(mktemp -d)
 
 FILE_1=`uuidgen`
 FILE_2=`uuidgen`
@@ -37,10 +36,13 @@ echo
 echo "Creating files: ${FILE_1} ${FILE_2}"
 
 FRONTEND_IP=$(kubectl -n ${NAMESPACE} get pods ctafrontend -o json | jq .status.podIP | tr -d '"')
+
+
+############## CREAT TEMPORARY DIRECTORY ##############
+TMP_DIR=$(mktemp -d)
 kubectl -n ${NAMESPACE} exec ctafrontend -- bash -c "mkdir -p ${TMP_DIR}"
 
-echo
-echo "ENABLE CTAFRONTEND TO EXECUTE CTA ADMIN COMMANDS"
+############## ENABLE CTAFRONTEND TO EXECUTE CTA ADMIN COMMANDS ##############
 kubectl --namespace=${NAMESPACE} exec kdc -- cat /root/ctaadmin2.keytab | kubectl --namespace=${NAMESPACE} exec -i ctafrontend --  bash -c "cat > /root/ctaadmin2.keytab; mkdir -p /tmp/ctaadmin2"
 kubectl -n ${NAMESPACE} cp client_helper.sh ctafrontend:/root/client_helper.sh
 touch ${TMP_DIR}/init_kerb.sh
@@ -48,37 +50,30 @@ echo '. /root/client_helper.sh; admin_kinit' >> ${TMP_DIR}/init_kerb.sh
 kubectl -n ${NAMESPACE} cp ${TMP_DIR}/init_kerb.sh ctafrontend:${TMP_DIR}/init_kerb.sh
 kubectl -n ${NAMESPACE} exec ctafrontend -- bash ${TMP_DIR}/init_kerb.sh
 
-
-echo
-echo "ADD FRONTEND GATEWAY TO EOS"
-echo "kubectl -n ${NAMESPACE} exec ctaeos -- bash eos root://${EOSINSTANCE} -r 0 0 vid add gateway ${FRONTEND_IP} grpc"
-kubectl -n ${NAMESPACE} exec ctaeos -- eos -r 0 0 vid add gateway ${FRONTEND_IP} grpc
-
-echo
-echo "ADD STORAGE CLASS WITH ONE COPIES ${NEW_STORAGE_CLASS_NAME}"
+############## ADD STORAGE CLASS WITH TWO COPIES ##############
 kubectl -n ${NAMESPACE} exec ctacli -- cta-admin sc add --name ${NEW_STORAGE_CLASS_NAME} --numberofcopies 2 --virtualorganisation vo --comment "comment"
 kubectl -n ${NAMESPACE} exec ctacli -- cta-admin sc ls
 
-echo
-echo "COPY REQUIRED FILES TO FRONTEND POD"
+############## COPY REQUIRED FILES TO FRONTEND POD ##############
 echo "kubectl cp ${NAMESPACE}/ctacli:/etc/cta/cta-cli.conf ${TMP_DIR}/cta-cli.conf"
 echo "kubectl cp ${TMP_DIR}/cta-cli.conf ${NAMESPACE}/ctafrontend:/etc/cta/cta-cli.conf"
 kubectl cp ${NAMESPACE}/ctacli:/etc/cta/cta-cli.conf ${TMP_DIR}/cta-cli.conf
 kubectl cp ${TMP_DIR}/cta-cli.conf ${NAMESPACE}/ctafrontend:/etc/cta/cta-cli.conf
 
+############## ARCHIVE FILES ##############
 kubectl -n ${NAMESPACE} exec client -- bash -c "mkdir -p ${TMP_DIR}"
 kubectl -n ${NAMESPACE} cp common/archive_file.sh client:${TMP_DIR}/
 kubectl -n ${NAMESPACE} cp client_helper.sh client:/root/
 kubectl -n ${NAMESPACE} exec client -- bash ${TMP_DIR}/archive_file.sh -f ${FILE_1} || exit 1
 kubectl -n ${NAMESPACE} exec client -- bash ${TMP_DIR}/archive_file.sh -f ${FILE_2} || exit 1
 
+############## GET EOS AND ARCHVIVE FILE IDS FROM ARCHIVED FILES ##############
 EOS_METADATA_PATH_1=$(mktemp -d).json
 echo "SEND EOS METADATA TO JSON FILE: ${EOS_METADATA_PATH}"
 touch ${EOS_METADATA_PATH_1}
 kubectl -n ${NAMESPACE} exec client -- eos -j root://${EOSINSTANCE} file info /eos/ctaeos/cta/${FILE_1} | jq . | tee ${EOS_METADATA_PATH_1}
 EOS_ARCHIVE_ID_1=$(jq -r '.xattr | .["sys.archive.file_id"]' ${EOS_METADATA_PATH_1})
 EOS_FILE_ID_1=$(jq -r '.id' ${EOS_METADATA_PATH_1})
-
 
 EOS_METADATA_PATH_2=$(mktemp -d).json
 echo "SEND EOS METADATA TO JSON FILE: ${EOS_METADATA_PATH_2}"
@@ -87,35 +82,17 @@ kubectl -n ${NAMESPACE} exec client -- eos -j root://${EOSINSTANCE} file info /e
 EOS_ARCHIVE_ID_2=$(jq -r '.xattr | .["sys.archive.file_id"]' ${EOS_METADATA_PATH_2})
 EOS_FILE_ID_2=$(jq -r '.id' ${EOS_METADATA_PATH_2})
 
-echo
-echo "CHANGE FILES WITH IDS"
-cd ~
+############## CREATE INPUT FILE TO USE FOR TOOL ##############
 IDS_FILEPATH=${TMP_DIR}/archiveFileIds.txt
 touch ${IDS_FILEPATH}
 echo '{"archiveId": '${EOS_ARCHIVE_ID_1}', "fid": '${EOS_FILE_ID_1}', "instance": "'${EOSINSTANCE}'"}' >> ${IDS_FILEPATH}
 echo '{"archiveId": '${EOS_ARCHIVE_ID_2}', "fid": '${EOS_FILE_ID_2}', "instance": "'${EOSINSTANCE}'"}' >> ${IDS_FILEPATH}
-sleep 1
 
-echo
-echo "kubectl cp ${IDS_FILEPATH} ${NAMESPACE}/ctafrontend:${IDS_FILEPATH}"
+############## RUN TOOL ##############
 kubectl cp ${IDS_FILEPATH} ${NAMESPACE}/ctafrontend:${IDS_FILEPATH}
-echo "kubectl -n ${NAMESPACE} exec ctafrontend -- bash -c XrdSecPROTOCOL=krb5 KRB5CCNAME=/tmp/ctaadmin2/krb5cc_0 cta-change-storage-class --storageclassname ${NEW_STORAGE_CLASS_NAME} --json ${IDS_FILEPATH}"
-kubectl -n ${NAMESPACE} exec ctafrontend -- bash -c "XrdSecPROTOCOL=krb5 KRB5CCNAME=/tmp/ctaadmin2/krb5cc_0 cta-change-storage-class --storageclassname ${NEW_STORAGE_CLASS_NAME} --json ${IDS_FILEPATH} -t 1"
+kubectl -n ${NAMESPACE} exec ctafrontend -- bash -c "XrdSecPROTOCOL=krb5 KRB5CCNAME=/tmp/ctaadmin2/krb5cc_0 cta-change-storage-class-in-catalogue --storageclassname ${NEW_STORAGE_CLASS_NAME} --json ${IDS_FILEPATH} -t 1"
 
-EOS_METADATA_PATH_AFTER_CHANGE_1=$(mktemp -d).json
-echo "SEND EOS METADATA TO JSON FILE: ${EOS_METADATA_PATH_AFTER_CHANGE_1}"
-touch ${EOS_METADATA_PATH_AFTER_CHANGE_1}
-kubectl -n ${NAMESPACE} exec client -- eos -j root://${EOSINSTANCE} file info /eos/ctaeos/cta/${FILE_1} | jq . | tee ${EOS_METADATA_PATH_AFTER_CHANGE_1}
-EOS_STORAGE_CLASS_1=$(jq -r '.xattr | .["sys.archive.storage_class"]' ${EOS_METADATA_PATH_AFTER_CHANGE_1})
-rm -r ${EOS_METADATA_PATH_AFTER_CHANGE_1}
-
-EOS_METADATA_PATH_AFTER_CHANGE_2=$(mktemp -d).json
-echo "SEND EOS METADATA TO JSON FILE: ${EOS_METADATA_PATH_AFTER_CHANGE_2}"
-touch ${EOS_METADATA_PATH_AFTER_CHANGE_2}
-kubectl -n ${NAMESPACE} exec client -- eos -j root://${EOSINSTANCE} file info /eos/ctaeos/cta/${FILE_2} | jq . | tee ${EOS_METADATA_PATH_AFTER_CHANGE_2}
-EOS_STORAGE_CLASS_2=$(jq -r '.xattr | .["sys.archive.storage_class"]' ${EOS_METADATA_PATH_AFTER_CHANGE_2})
-rm -r ${EOS_METADATA_PATH_AFTER_CHANGE_2}
-
+############## CHECK THAT VALUES WERE CHANGED IN THE CATALOGUE ##############
 CATALOGUE_METADATA_PATH_AFTER_CHANGE_1=$(mktemp -d).json
 echo "SEND CATALOGUE METADATA TO JSON FILE: ${CATALOGUE_METADATA_PATH_AFTER_CHANGE_1}"
 touch ${CATALOGUE_METADATA_PATH_AFTER_CHANGE_1}
@@ -130,18 +107,8 @@ kubectl -n ${NAMESPACE} exec ctacli -- cta-admin --json tf ls --id ${EOS_ARCHIVE
 CATALOGUE_STORAGE_CLASS_2=$(jq . ${CATALOGUE_METADATA_PATH_AFTER_CHANGE_2} | jq '.[0]' | jq -r '.af | .["storageClass"]')
 rm -r ${CATALOGUE_METADATA_PATH_AFTER_CHANGE_2}
 
-if test ${EOS_STORAGE_CLASS_1} != ${NEW_STORAGE_CLASS_NAME}; then
-  echo "ERROR: File ${FILE_1} did not change the storage class in EOS"
-  exit 1
-fi
-
 if test ${CATALOGUE_STORAGE_CLASS_1} != ${NEW_STORAGE_CLASS_NAME}; then
   echo "ERROR: File ${FILE_1} did not change the storage class in the Catalogue"
-  exit 1
-fi
-
-if test ${EOS_STORAGE_CLASS_2} != ${NEW_STORAGE_CLASS_NAME}; then
-  echo "ERROR: File ${FILE_2} did not change the storage class in EOS"
   exit 1
 fi